Added encrypted export
